Deleted the old Tecla settings activity.
